Decouple gitea user id and git-ticket user name
This change makes use of the gitea user ID and stores it internally just
like the Phabricator ID for each user.

This way, the current restriction of using the same user name for gitea
and git ticket is removed.

When a user is created it may be created with
```
git ticket user create --gitea-username "Javier Alvarez Garcia"
```

If the command-line argument is not given and `--skip-gitea-id` is not
passed, git ticket will prompt the user during runtime.

Note that existing users will return ID -1, and all users must be
updated to fetch the gitea user IDs. This is a one-time operation.