Additional checks for CCB approval and block
Only allow approval and block of a ticket that is in a state that can
transition to the target state being approved/blocked.

This should prevent unintentional issues by which a CCB user would set the
approval of the accepted state, while they meant to set only the vetted
state.

Additionally, if possible, attempt to automatically transition to the target state,
assuming that the CCB validation passes (when there was only 1 CCB
member left to approve a state).