Apply missing change in index.js
This was somehow lost in the add comment PR and was causing set status
to fail. I also added a check for the index handler, because it also
get's matched with any path that has prefix "/". Instead of returning the
index when something does not match a path, let's return an error
message.