Fix review IDs in webui
There has been an unintended display format change in git ticket due to
recent change in https://gitea.daedalean.ai/daedalean-github/git-ticket/pulls/45

This change was only meant to affect the show command, but by changing
the behavior of the Id() function in the review interface of gitea, it
affected the webui as well, where it is not needed, as the `xref`
patterns already apply the right transformation to a link.

To reduce the possible side-effects of this change, let's just keep the
behavior of the old Id() method to return the gitea reference, while
adding a new one to actually get the ReviewUrl(). Use the new method in
the show command instead of Id().

In addition, implement this method for Phabricator reviews as well,
given there is no reason for divergence in their behavior.