Add flag to force CCB action (approval/block)
Sometimes the scope of a ticket is modified while being worked on, due
to new information being found. At this stage, we can add new CCB
members for vetted, but the current workflow doesn't allow it.

I suggest we add a flag to force it so that we don't have to transition
the ticket to rejected and then proposed, and can instead just add it in
the current state.