Fix checklists links
In a previous change I already attempted to fix checklists because the
matcher had the trailing `/`, but the link did not. I attempted to fix
it by removing the trailing `/` and even though it looks like it worked
for most checklists, there were some links that are being translated by
the browser to contain the trailing `/`, now making it fail.

Instead, let's use the trailing `/` always and make sure that the link
we generate in the tickets folder has the trailing `/`.